Migration of the csproj format with some namespaces refactoring plus tests project.
