+++ MacroProcessing.Sandbox/MacroProcessing.Sandbox.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net461</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\MacroProcessing\MacroProcessing.csproj" />
+  </ItemGroup>
+</Project>
