+++ MacroProcessing.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26228.10
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MacroProcessing", "MacroProcessing\MacroProcessing.csproj", "{18B9A7AF-A2DD-479E-B645-E3D08FCD6E94}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MacroProcessing.Sandbox", "MacroProcessing.Sandbox\MacroProcessing.Sandbox.csproj", "{40A86AFA-BF79-4BE4-A40A-3ED85DAF0E96}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MacroProcessing.Tests", "MacroProcessing.Tests\MacroProcessing.Tests.csproj", "{72254420-4019-461D-A30E-06667C1FB779}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{18B9A7AF-A2DD-479E-B645-E3D08FCD6E94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{18B9A7AF-A2DD-479E-B645-E3D08FCD6E94}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{18B9A7AF-A2DD-479E-B645-E3D08FCD6E94}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{18B9A7AF-A2DD-479E-B645-E3D08FCD6E94}.Release|Any CPU.Build.0 = Release|Any CPU
+		{40A86AFA-BF79-4BE4-A40A-3ED85DAF0E96}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{40A86AFA-BF79-4BE4-A40A-3ED85DAF0E96}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{40A86AFA-BF79-4BE4-A40A-3ED85DAF0E96}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{40A86AFA-BF79-4BE4-A40A-3ED85DAF0E96}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72254420-4019-461D-A30E-06667C1FB779}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72254420-4019-461D-A30E-06667C1FB779}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72254420-4019-461D-A30E-06667C1FB779}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72254420-4019-461D-A30E-06667C1FB779}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5D8F94CF-1A32-41CC-8BFA-EBFBD9FD4932}
+	EndGlobalSection
+EndGlobal
