+++ MacroProcessing/CompiledTemplate.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace SfySoft.MacroProcessing.Core
+namespace MacroProcessing
 {
     public class CompiledTemplate
     {
