+++ MacroProcessing/KeywordRegistry.cs
@@ -3,10 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace SfySoft.MacroProcessing.Core
+namespace MacroProcessing
 {
     public class KeywordRegistry
     {
