+++ MacroProcessing/MacroProcessor.cs
@@ -1,12 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
-using System.Text;
 
-namespace SfySoft.MacroProcessing.Core
+namespace MacroProcessing
 {
     public static class MacroProcessor
     {
