--- SfySoft.MacroProcessing.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26228.10
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SfySoft.MacroProcessing.Core", "SfySoft.MacroProcessing.Core\SfySoft.MacroProcessing.Core.csproj", "{5D1E1661-5FDB-48DF-868A-20B09509047C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SfySoft.MacroProcessing.Sandbox", "SfySoft.MacroProcessing.Sandbox\SfySoft.MacroProcessing.Sandbox.csproj", "{EE89F1E7-8FFF-4AF4-8B52-CCF57471CDB4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5D1E1661-5FDB-48DF-868A-20B09509047C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5D1E1661-5FDB-48DF-868A-20B09509047C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5D1E1661-5FDB-48DF-868A-20B09509047C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5D1E1661-5FDB-48DF-868A-20B09509047C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EE89F1E7-8FFF-4AF4-8B52-CCF57471CDB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE89F1E7-8FFF-4AF4-8B52-CCF57471CDB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE89F1E7-8FFF-4AF4-8B52-CCF57471CDB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE89F1E7-8FFF-4AF4-8B52-CCF57471CDB4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
